requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0
weasyprint>=61.0
